+++ backend/src/main/java/com/eventra/StatusController.java
@@ -1,4 +1,4 @@
-package com.eventra.controller;
+package com.eventra;
 
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RestController;
